split tablegen helper into .h and .cpp
